Split ControlPanel into its eleven sections
The panel was 578 lines in one function, already divided by banner comments into
Canvas, Style, Text, Text Style, Colors, Shape, Logo, Timing, Position, Actions
and Library. Each is now a composable of its own in ControlPanelSections.kt, and
ControlPanel is the scroll container plus eleven calls.

Text Style was long enough on its own to split again, into the font/size, weight
and case rows.

One real fix fell out of it: FieldRow's content lambda had no RowScope, so every
`Modifier.weight` inside a field row was resolving against the enclosing Column
instead of the Row that measures it. The lambda now carries the RowScope.

Styles dumped before and after: unchanged.
