+++ lottieGenerator/src/main/kotlin/org/churchpresenter/lottiegen/ui/ControlPanel.kt
@@ -1,8 +1,8 @@
 package org.churchpresenter.lottiegen.ui
 
+import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -18,10 +18,7 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.ExposedDropdownMenuAnchorType
-import androidx.compose.material3.ExposedDropdownMenuBox
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -36,43 +33,28 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import org.churchpresenter.lottiegen.LottieGenState
-import org.churchpresenter.lottiegen.model.CANVAS_PRESETS
-import org.churchpresenter.lottiegen.model.LottieFont
-import org.churchpresenter.lottiegen.model.StyleCatalog
-import org.churchpresenter.lottiegen.model.TIMING_PRESETS
 import org.churchpresenter.lottiegen.ui.components.AccentButton
-import org.churchpresenter.lottiegen.ui.components.CollapsibleSection
-import org.churchpresenter.lottiegen.ui.components.ColorPickerRow
-import org.churchpresenter.lottiegen.ui.components.DeleteIconButton
-import org.churchpresenter.lottiegen.ui.components.LottieCheckbox
-import org.churchpresenter.lottiegen.ui.components.LottieDropdown
 import org.churchpresenter.lottiegen.ui.components.LottieTextField
-import org.churchpresenter.lottiegen.ui.components.SectionCard
-import org.churchpresenter.lottiegen.ui.components.SegmentedButtons
-import org.churchpresenter.lottiegen.ui.components.SliderWithLabel
 import org.churchpresenter.lottiegen.ui.components.SubtleButton
-import javax.swing.JFileChooser
-import javax.swing.SwingUtilities
-import javax.swing.filechooser.FileNameExtensionFilter
 
 /** The ranges every numeric field in this panel is clamped to. */
-private const val MIN_CANVAS_PX = 100
-private const val MAX_CANVAS_W_PX = 7680
-private const val MAX_CANVAS_H_PX = 4320
-private const val MIN_BASE_SIZE = 10
-private const val MAX_BASE_SIZE = 80
-private const val MIN_TEXT_EM = 0.5f
-private const val MAX_TEXT_EM = 4f
-private const val MIN_ANIM_SECONDS = 0.5f
-private const val MAX_ANIM_SECONDS = 20f
-private const val MAX_HOLD_SECONDS = 30f
-private const val MIN_NUDGE_EM = -0.5f
-private const val MAX_NUDGE_EM = 1f
+internal const val MIN_CANVAS_PX = 100
+internal const val MAX_CANVAS_W_PX = 7680
+internal const val MAX_CANVAS_H_PX = 4320
+internal const val MIN_BASE_SIZE = 10
+internal const val MAX_BASE_SIZE = 80
+internal const val MIN_TEXT_EM = 0.5f
+internal const val MAX_TEXT_EM = 4f
+internal const val MIN_ANIM_SECONDS = 0.5f
+internal const val MAX_ANIM_SECONDS = 20f
+internal const val MAX_HOLD_SECONDS = 30f
+internal const val MIN_NUDGE_EM = -0.5f
+internal const val MAX_NUDGE_EM = 1f
 
 
 /** Two equal columns, the layout every field pair in the panel uses. */
 @Composable
-private fun FieldRow(content: @Composable () -> Unit) {
+internal fun FieldRow(content: @Composable RowScope.() -> Unit) {
     Row(
         modifier = Modifier.fillMaxWidth(),
         horizontalArrangement = Arrangement.spacedBy(7.dp)
@@ -130,7 +112,6 @@ private fun PanelHeader() {
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun ControlPanel(viewModel: LottieGenState, panelWidth: Dp = 436.dp) {
-    val cfg = viewModel.config
     val scrollState = rememberScrollState()
     var showBatchImport by remember { mutableStateOf(false) }
     var batchImportText by remember { mutableStateOf("") }
@@ -153,519 +134,17 @@ fun ControlPanel(viewModel: LottieGenState, panelWidth: Dp = 436.dp) {
             verticalArrangement = Arrangement.spacedBy(7.dp)
         ) {
 
-            // ═══ Canvas ═══
-            CollapsibleSection(Strings.sectionCanvas, hint = "${cfg.canvasW} × ${cfg.canvasH}") {
-                val canvasIndex = CANVAS_PRESETS.indexOfFirst {
-                    it.width == cfg.canvasW && it.height == cfg.canvasH
-                }
-                SegmentedButtons(
-                    labels = CANVAS_PRESETS.map { it.label },
-                    selectedIndex = canvasIndex,
-                    onSelect = { i ->
-                        val p = CANVAS_PRESETS[i]
-                        viewModel.updateConfig { it.copy(canvasW = p.width, canvasH = p.height) }
-                    }
-                )
-                FieldRow {
-                    LottieTextField(
-                        value = cfg.canvasW.toString(),
-                        onValueChange = { v ->
-                            v.toIntOrNull()?.let {
-                                viewModel.updateConfig { c ->
-                                    c.copy(canvasW = it.coerceIn(MIN_CANVAS_PX, MAX_CANVAS_W_PX))
-                                }
-                            }
-                        },
-                        label = Strings.width,
-                        modifier = Modifier.weight(1f), fillWidth = true, singleLine = true
-                    )
-                    LottieTextField(
-                        value = cfg.canvasH.toString(),
-                        onValueChange = { v ->
-                            v.toIntOrNull()?.let {
-                                viewModel.updateConfig { c ->
-                                    c.copy(canvasH = it.coerceIn(MIN_CANVAS_PX, MAX_CANVAS_H_PX))
-                                }
-                            }
-                        },
-                        label = Strings.height,
-                        modifier = Modifier.weight(1f), fillWidth = true, singleLine = true
-                    )
-                }
-                SliderWithLabel(
-                    Strings.scale, cfg.scaleFactor,
-                    { viewModel.updateConfig { c -> c.copy(scaleFactor = it) } },
-                    0.5f..3f, unit = "×", format = { "%.2f".format(it) }
-                )
-            }
-
-            // ═══ Style & Layout ═══
-            SectionCard(Strings.sectionStyleLayout) {
-                FieldRow {
-                    var styleExpanded by remember { mutableStateOf(false) }
-                    ExposedDropdownMenuBox(styleExpanded, { styleExpanded = it }, Modifier.weight(1f)) {
-                        LottieDropdown(
-                            label = Strings.style,
-                            value = StyleCatalog.labelFor(cfg.style),
-                            expanded = styleExpanded,
-                            modifier = Modifier.fillMaxWidth()
-                                .menuAnchor(ExposedDropdownMenuAnchorType.PrimaryNotEditable)
-                        )
-                        ExposedDropdownMenu(styleExpanded, { styleExpanded = false }) {
-                            StyleCatalog.entries.forEach { style ->
-                                DropdownMenuItem(
-                                    text = { Text(style.label) },
-                                    onClick = {
-                                        viewModel.updateConfig { it.copy(style = style.id) }
-                                        styleExpanded = false
-                                    }
-                                )
-                            }
-                        }
-                    }
-                    var alignExpanded by remember { mutableStateOf(false) }
-                    val alignmentLabels = mapOf(
-                        "left" to Strings.alignLeft,
-                        "center" to Strings.alignCenter,
-                        "right" to Strings.alignRight
-                    )
-                    ExposedDropdownMenuBox(alignExpanded, { alignExpanded = it }, Modifier.weight(1f)) {
-                        LottieDropdown(
-                            label = Strings.alignment,
-                            value = alignmentLabels[cfg.align] ?: cfg.align,
-                            expanded = alignExpanded,
-                            modifier = Modifier.fillMaxWidth()
-                                .menuAnchor(ExposedDropdownMenuAnchorType.PrimaryNotEditable)
-                        )
-                        ExposedDropdownMenu(alignExpanded, { alignExpanded = false }) {
-                            alignmentLabels.forEach { (id, label) ->
-                                DropdownMenuItem(
-                                    text = { Text(label) },
-                                    onClick = { viewModel.updateConfig { it.copy(align = id) }; alignExpanded = false }
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-
-            // ═══ Text ═══
-            SectionCard(Strings.sectionText) {
-                LottieTextField(
-                    value = cfg.nameText,
-                    onValueChange = { viewModel.updateConfig { c -> c.copy(nameText = it) } },
-                    label = Strings.name,
-                    modifier = Modifier.fillMaxWidth(), fillWidth = true, singleLine = true
-                )
-                LottieTextField(
-                    value = cfg.infoText,
-                    onValueChange = { viewModel.updateConfig { c -> c.copy(infoText = it) } },
-                    label = Strings.info,
-                    modifier = Modifier.fillMaxWidth(), fillWidth = true, singleLine = true
-                )
-            }
-
-            // ═══ Text Style ═══
-            CollapsibleSection(Strings.sectionTextStyle, hint = cfg.fontFamily) {
-                FieldRow {
-                    var fontExpanded by remember { mutableStateOf(false) }
-                    ExposedDropdownMenuBox(fontExpanded, { fontExpanded = it }, Modifier.weight(1f)) {
-                        LottieDropdown(
-                            label = Strings.font,
-                            value = cfg.fontFamily,
-                            expanded = fontExpanded,
-                            modifier = Modifier.fillMaxWidth()
-                                .menuAnchor(ExposedDropdownMenuAnchorType.PrimaryNotEditable)
-                        )
-                        ExposedDropdownMenu(fontExpanded, { fontExpanded = false }) {
-                            LottieFont.entries.forEach { font ->
-                                DropdownMenuItem(
-                                    text = { Text(font.familyName) },
-                                    onClick = {
-                                        viewModel.updateConfig { it.copy(fontFamily = font.familyName) }
-                                        fontExpanded = false
-                                    }
-                                )
-                            }
-                        }
-                    }
-                    LottieTextField(
-                        value = cfg.baseSize.toString(),
-                        onValueChange = { v ->
-                            v.toIntOrNull()?.let {
-                                viewModel.updateConfig { c ->
-                                    c.copy(baseSize = it.coerceIn(MIN_BASE_SIZE, MAX_BASE_SIZE))
-                                }
-                            }
-                        },
-                        label = Strings.baseSize,
-                        modifier = Modifier.weight(1f), fillWidth = true, singleLine = true
-                    )
-                }
-                FieldRow {
-                    LottieTextField(
-                        value = cfg.nameSize.toString(),
-                        onValueChange = { v ->
-                            v.toFloatOrNull()?.let {
-                                viewModel.updateConfig { c -> c.copy(nameSize = it.coerceIn(MIN_TEXT_EM, MAX_TEXT_EM)) }
-                            }
-                        },
-                        label = Strings.nameSize, modifier = Modifier.weight(1f), fillWidth = true, singleLine = true
-                    )
-                    LottieTextField(
-                        value = cfg.infoSize.toString(),
-                        onValueChange = { v ->
-                            v.toFloatOrNull()?.let {
-                                viewModel.updateConfig { c -> c.copy(infoSize = it.coerceIn(MIN_TEXT_EM, MAX_TEXT_EM)) }
-                            }
-                        },
-                        label = Strings.infoSize, modifier = Modifier.weight(1f), fillWidth = true, singleLine = true
-                    )
-                }
-                FieldRow {
-                    var nwExpanded by remember { mutableStateOf(false) }
-                    ExposedDropdownMenuBox(nwExpanded, { nwExpanded = it }, Modifier.weight(1f)) {
-                        LottieDropdown(
-                            label = Strings.nameWeight,
-                            value = if (cfg.nameWeight >= 700) Strings.bold else Strings.normal,
-                            expanded = nwExpanded,
-                            modifier = Modifier.fillMaxWidth()
-                                .menuAnchor(ExposedDropdownMenuAnchorType.PrimaryNotEditable)
-                        )
-                        ExposedDropdownMenu(nwExpanded, { nwExpanded = false }) {
-                            DropdownMenuItem(
-                                { Text(Strings.bold) },
-                                { viewModel.updateConfig { it.copy(nameWeight = 700) }; nwExpanded = false },
-                            )
-                            DropdownMenuItem(
-                                { Text(Strings.normal) },
-                                { viewModel.updateConfig { it.copy(nameWeight = 400) }; nwExpanded = false },
-                            )
-                        }
-                    }
-                    var iwExpanded by remember { mutableStateOf(false) }
-                    ExposedDropdownMenuBox(iwExpanded, { iwExpanded = it }, Modifier.weight(1f)) {
-                        LottieDropdown(
-                            label = Strings.infoWeight,
-                            value = if (cfg.infoWeight >= 700) Strings.bold else Strings.normal,
-                            expanded = iwExpanded,
-                            modifier = Modifier.fillMaxWidth()
-                                .menuAnchor(ExposedDropdownMenuAnchorType.PrimaryNotEditable)
-                        )
-                        ExposedDropdownMenu(iwExpanded, { iwExpanded = false }) {
-                            DropdownMenuItem(
-                                { Text(Strings.normal) },
-                                { viewModel.updateConfig { it.copy(infoWeight = 400) }; iwExpanded = false },
-                            )
-                            DropdownMenuItem(
-                                { Text(Strings.bold) },
-                                { viewModel.updateConfig { it.copy(infoWeight = 700) }; iwExpanded = false },
-                            )
-                        }
-                    }
-                }
-                FieldRow {
-                    var ntExpanded by remember { mutableStateOf(false) }
-                    ExposedDropdownMenuBox(ntExpanded, { ntExpanded = it }, Modifier.weight(1f)) {
-                        LottieDropdown(
-                            label = Strings.nameTransform,
-                            value = if (cfg.nameTransform == "uppercase") Strings.uppercase else Strings.none,
-                            expanded = ntExpanded,
-                            modifier = Modifier.fillMaxWidth()
-                                .menuAnchor(ExposedDropdownMenuAnchorType.PrimaryNotEditable)
-                        )
-                        ExposedDropdownMenu(ntExpanded, { ntExpanded = false }) {
-                            DropdownMenuItem(
-                                { Text(Strings.uppercase) },
-                                { viewModel.updateConfig { it.copy(nameTransform = "uppercase") }; ntExpanded = false },
-                            )
-                            DropdownMenuItem(
-                                { Text(Strings.none) },
-                                { viewModel.updateConfig { it.copy(nameTransform = "none") }; ntExpanded = false },
-                            )
-                        }
-                    }
-                    var itExpanded by remember { mutableStateOf(false) }
-                    ExposedDropdownMenuBox(itExpanded, { itExpanded = it }, Modifier.weight(1f)) {
-                        LottieDropdown(
-                            label = Strings.infoTransform,
-                            value = if (cfg.infoTransform == "uppercase") Strings.uppercase else Strings.none,
-                            expanded = itExpanded,
-                            modifier = Modifier.fillMaxWidth()
-                                .menuAnchor(ExposedDropdownMenuAnchorType.PrimaryNotEditable)
-                        )
-                        ExposedDropdownMenu(itExpanded, { itExpanded = false }) {
-                            DropdownMenuItem(
-                                { Text(Strings.none) },
-                                { viewModel.updateConfig { it.copy(infoTransform = "none") }; itExpanded = false },
-                            )
-                            DropdownMenuItem(
-                                { Text(Strings.uppercase) },
-                                { viewModel.updateConfig { it.copy(infoTransform = "uppercase") }; itExpanded = false },
-                            )
-                        }
-                    }
-                }
-            }
-
-            // ═══ Colors ═══
-            CollapsibleSection(Strings.sectionColors, initiallyExpanded = true) {
-                Column(verticalArrangement = Arrangement.spacedBy(10.dp)) {
-                    ColorPickerRow(Strings.colorNameText, cfg.nameColor, cfg.nameColorAlpha,
-                        { viewModel.updateConfig { c -> c.copy(nameColor = it) } },
-                        { viewModel.updateConfig { c -> c.copy(nameColorAlpha = it) } })
-                    ColorPickerRow(Strings.colorInfoText, cfg.infoColor, cfg.infoColorAlpha,
-                        { viewModel.updateConfig { c -> c.copy(infoColor = it) } },
-                        { viewModel.updateConfig { c -> c.copy(infoColorAlpha = it) } })
-                    ColorPickerRow(Strings.colorAccent, cfg.accentColor, cfg.accentColorAlpha,
-                        { viewModel.updateConfig { c -> c.copy(accentColor = it) } },
-                        { viewModel.updateConfig { c -> c.copy(accentColorAlpha = it) } })
-                    ColorPickerRow(Strings.colorBackground, cfg.bgColor, cfg.bgColorAlpha,
-                        { viewModel.updateConfig { c -> c.copy(bgColor = it) } },
-                        { viewModel.updateConfig { c -> c.copy(bgColorAlpha = it) } })
-                    ColorPickerRow(Strings.colorBorder, cfg.borderColor, cfg.borderColorAlpha,
-                        { viewModel.updateConfig { c -> c.copy(borderColor = it) } },
-                        { viewModel.updateConfig { c -> c.copy(borderColorAlpha = it) } })
-                }
-
-                SubtleButton(Strings.saveColors, { viewModel.saveColorTheme() })
-
-                viewModel.colorThemes.forEachIndexed { i, theme ->
-                    Row(
-                        modifier = Modifier.fillMaxWidth().clickable { viewModel.loadColorTheme(i) },
-                        verticalAlignment = Alignment.CenterVertically,
-                        horizontalArrangement = Arrangement.spacedBy(8.dp)
-                    ) {
-                        Text(
-                            theme.name,
-                            modifier = Modifier.weight(1f),
-                            fontSize = 12.sp,
-                            color = Tokens.LabelText,
-                            maxLines = 1,
-                        )
-                        DeleteIconButton({ viewModel.deleteColorTheme(i) })
-                    }
-                }
-            }
-
-            // ═══ Shape ═══
-            CollapsibleSection(Strings.sectionShape) {
-                SliderWithLabel(
-                    Strings.corners, cfg.corners,
-                    { viewModel.updateConfig { c -> c.copy(corners = it) } },
-                    0f..2f, unit = "em", format = { "%.2f".format(it) }
-                )
-                SliderWithLabel(
-                    Strings.borderThickness, cfg.borderThickness,
-                    { viewModel.updateConfig { c -> c.copy(borderThickness = it) } },
-                    0f..5f, format = { "%.2f".format(it) }
-                )
-                Column(verticalArrangement = Arrangement.spacedBy(9.dp)) {
-                    FieldRow {
-                        LottieCheckbox(Strings.showBackground, cfg.bgEnabled,
-                            { viewModel.updateConfig { c -> c.copy(bgEnabled = it) } }, Modifier.weight(1f))
-                        LottieCheckbox(Strings.shadow, cfg.shadowEnabled,
-                            { viewModel.updateConfig { c -> c.copy(shadowEnabled = it) } }, Modifier.weight(1f))
-                    }
-                    FieldRow {
-                        LottieCheckbox(Strings.hideName, cfg.hideName,
-                            { viewModel.updateConfig { c -> c.copy(hideName = it) } }, Modifier.weight(1f))
-                        LottieCheckbox(Strings.hideInfo, cfg.hideInfo,
-                            { viewModel.updateConfig { c -> c.copy(hideInfo = it) } }, Modifier.weight(1f))
-                    }
-                }
-            }
-
-            // ═══ Logo ═══
-            CollapsibleSection(
-                Strings.sectionLogo,
-                hint = if (cfg.logoEnabled) cfg.logoSelect.ifEmpty { Strings.logoNone } else Strings.logoNone
-            ) {
-                LottieCheckbox(Strings.showLogo, cfg.logoEnabled,
-                    { viewModel.updateConfig { c -> c.copy(logoEnabled = it) } })
-
-                var logoExpanded by remember { mutableStateOf(false) }
-                ExposedDropdownMenuBox(logoExpanded, { logoExpanded = it }, Modifier.fillMaxWidth()) {
-                    LottieDropdown(
-                        label = Strings.logoLabel,
-                        value = cfg.logoSelect.ifEmpty { Strings.logoNone },
-                        expanded = logoExpanded,
-                        modifier = Modifier.fillMaxWidth().menuAnchor(ExposedDropdownMenuAnchorType.PrimaryNotEditable)
-                    )
-                    ExposedDropdownMenu(logoExpanded, { logoExpanded = false }) {
-                        DropdownMenuItem({ Text(Strings.logoNone) }, { viewModel.clearLogo(); logoExpanded = false })
-                        viewModel.availableLogos.forEach { logo ->
-                            DropdownMenuItem({ Text(logo) }, { viewModel.selectLogo(logo); logoExpanded = false })
-                        }
-                        DropdownMenuItem({ Text(Strings.logoImport) }, {
-                            logoExpanded = false
-                            SwingUtilities.invokeLater {
-                                val chooser = JFileChooser()
-                                chooser.fileFilter = FileNameExtensionFilter(
-                                    "Images",
-                                    "png",
-                                    "jpg",
-                                    "jpeg",
-                                    "svg",
-                                    "webp",
-                                )
-                                if (chooser.showOpenDialog(null) == JFileChooser.APPROVE_OPTION) {
-                                    viewModel.importAndLoadLogo(chooser.selectedFile)
-                                }
-                            }
-                        })
-                    }
-                }
-
-                SliderWithLabel(
-                    Strings.logoSize, cfg.logoSize,
-                    { viewModel.updateConfig { c -> c.copy(logoSize = it) } },
-                    2f..6f, unit = "em"
-                )
-            }
-
-            // ═══ Timing ═══
-            CollapsibleSection(
-                Strings.sectionTiming,
-                hint = "%.1fs".format(cfg.animDuration + cfg.holdDuration)
-            ) {
-                val timingIndex = TIMING_PRESETS.indexOfFirst {
-                    it.animDuration == cfg.animDuration && it.holdDuration == cfg.holdDuration
-                }
-                SegmentedButtons(
-                    labels = TIMING_PRESETS.map { it.label },
-                    selectedIndex = timingIndex,
-                    onSelect = { i ->
-                        val p = TIMING_PRESETS[i]
-                        viewModel.updateConfig { it.copy(animDuration = p.animDuration, holdDuration = p.holdDuration) }
-                    }
-                )
-                FieldRow {
-                    LottieTextField(
-                        value = cfg.animDuration.toString(),
-                        onValueChange = { v ->
-                            v.toFloatOrNull()?.let {
-                                viewModel.updateConfig { c ->
-                                    c.copy(animDuration = it.coerceIn(MIN_ANIM_SECONDS, MAX_ANIM_SECONDS))
-                                }
-                            }
-                        },
-                        label = Strings.animDuration, modifier = Modifier.weight(1f),
-                        fillWidth = true, singleLine = true
-                    )
-                    LottieTextField(
-                        value = cfg.holdDuration.toString(),
-                        onValueChange = { v ->
-                            v.toFloatOrNull()?.let {
-                                viewModel.updateConfig { c -> c.copy(holdDuration = it.coerceIn(0f, MAX_HOLD_SECONDS)) }
-                            }
-                        },
-                        label = Strings.holdDuration, modifier = Modifier.weight(1f),
-                        fillWidth = true, singleLine = true
-                    )
-                }
-            }
-
-            // ═══ Position ═══
-            CollapsibleSection(Strings.sectionPosition) {
-                SliderWithLabel(
-                    Strings.marginH, cfg.marginH,
-                    { viewModel.updateConfig { c -> c.copy(marginH = it) } },
-                    0f..10f, unit = "rem"
-                )
-                SliderWithLabel(
-                    Strings.marginV, cfg.marginV,
-                    { viewModel.updateConfig { c -> c.copy(marginV = it) } },
-                    0f..10f, unit = "rem"
-                )
-                SliderWithLabel(
-                    Strings.lineSpacing, cfg.lineSpacing,
-                    { viewModel.updateConfig { c -> c.copy(lineSpacing = it) } },
-                    MIN_NUDGE_EM..MAX_NUDGE_EM, unit = "em", format = { "%.2f".format(it) }
-                )
-            }
-
-            // ═══ Actions ═══
-            Spacer(Modifier.height(5.dp))
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                if (viewModel.hasOutputDir) {
-                    AccentButton(Strings.saveLowerThird, { viewModel.saveLowerThird() }, Modifier.weight(1f))
-                } else {
-                    AccentButton(
-                        Strings.downloadJson,
-                        {
-                            SwingUtilities.invokeLater {
-                                val chooser = JFileChooser()
-                                chooser.fileSelectionMode = JFileChooser.DIRECTORIES_ONLY
-                                if (chooser.showOpenDialog(null) == JFileChooser.APPROVE_OPTION) {
-                                    viewModel.downloadJson(chooser.selectedFile)
-                                }
-                            }
-                        },
-                        Modifier.weight(1f)
-                    )
-                    SubtleButton(Strings.saveToLibrary, { viewModel.savePreset() }, Modifier.weight(1f))
-                }
-            }
-
-            // ═══ Library ═══
-            SectionCard(
-                Strings.sectionLibrary,
-                modifier = Modifier.padding(top = 3.dp),
-                trailing = { SubtleButton(Strings.batchImport, { showBatchImport = true }, compact = true) }
-            ) {
-                if (viewModel.presets.isNotEmpty()) {
-                    Row(horizontalArrangement = Arrangement.spacedBy(6.dp)) {
-                        SubtleButton(Strings.applyStyleToAll, { viewModel.applyStyleToAll() }, compact = true)
-                        if (viewModel.hasOutputDir) {
-                            SubtleButton(
-                                Strings.saveAllLowerThirds,
-                                { viewModel.batchDownloadAll(null) },
-                                compact = true,
-                            )
-                        } else {
-                            SubtleButton(Strings.saveAllLowerThirds, {
-                                SwingUtilities.invokeLater {
-                                    val chooser = JFileChooser()
-                                    chooser.fileSelectionMode = JFileChooser.DIRECTORIES_ONLY
-                                    if (chooser.showOpenDialog(null) == JFileChooser.APPROVE_OPTION) {
-                                        viewModel.batchDownloadAll(chooser.selectedFile)
-                                    }
-                                }
-                            }, compact = true)
-                        }
-                    }
-                }
-
-                if (viewModel.presets.isEmpty()) {
-                    Box(Modifier.fillMaxWidth().padding(vertical = 7.dp), contentAlignment = Alignment.Center) {
-                        Text(Strings.noPresets, fontSize = 12.sp, color = Tokens.UnitText)
-                    }
-                } else {
-                    viewModel.presets.forEachIndexed { i, preset ->
-                        Row(
-                            modifier = Modifier.fillMaxWidth().clickable { viewModel.loadPreset(i) },
-                            verticalAlignment = Alignment.CenterVertically,
-                            horizontalArrangement = Arrangement.spacedBy(8.dp)
-                        ) {
-                            Column(modifier = Modifier.weight(1f)) {
-                                Text(preset.name, fontSize = 12.sp, color = Tokens.PrimaryText, maxLines = 1)
-                                Text(
-                                    "${preset.config.canvasW}×${preset.config.canvasH} · " +
-                                        StyleCatalog.labelFor(preset.config.style),
-                                    fontSize = 10.sp,
-                                    color = Tokens.UnitText,
-                                    maxLines = 1
-                                )
-                            }
-                            DeleteIconButton({ viewModel.deletePreset(i) })
-                        }
-                    }
-                }
-            }
+            CanvasSection(viewModel)
+            StyleLayoutSection(viewModel)
+            TextSection(viewModel)
+            TextStyleSection(viewModel)
+            ColorsSection(viewModel)
+            ShapeSection(viewModel)
+            LogoSection(viewModel)
+            TimingSection(viewModel)
+            PositionSection(viewModel)
+            ActionsSection(viewModel)
+            LibrarySection(viewModel, onBatchImport = { showBatchImport = true })
         }
     }
